GPII-2177: Rename NexusConfig.js to CoOccurrenceEngine.js
