+++ index.js
@@ -16,4 +16,4 @@ var fluid = require("infusion");
 
 fluid.module.register("gpii-nexus", __dirname, require);
 require("./src/Nexus.js");
-require("./src/NexusConfig.js");
+require("./src/CoOccurrenceEngine.js");
